function_app/requirements.txt
azure-functions>=1.18.0
azure-identity>=1.15.0
pyodbc>=5.0.0
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
msal>=1.24.0
requests>=2.31.0
